dist/distrib: generate vi.0
The file is absent from the repository, not generated with "make",
but attempted to be installed with "make install".

An alternative to changing directories directly would be to use
"make -C", but "make -C" is not POSIX.  Is is present on Linux and BSDs,
but not on Solaris, for example.

Signed-off-by: Felix Janda <felix.janda@posteo.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>